Move part of asyncRename logic into MetastoreUtil
Moved logic:
- getFileSystem
- renameFile
- waitForAsyncRename
